Add cli functionality for config management

+++ src/python/dynami/dynami.py
@@ -7,7 +7,7 @@ from pathlib import Path
 # ConfigControl is a class to manage the config file
 from provider import *
 from ext.client import Client
-from ext.utils import PublicIP, ConfigControl
+from ext.utils import PublicIP
 
 available_providers = [
     "hetzner",
@@ -24,31 +24,34 @@ available_record_types = [
     "SRV"
 ]
 
-def check_config_path(base: Path) -> None:
-    # If path is not existing, create it
-    path_list = ["config", "cache"]
-    if not os.path.exists(base):
-        os.mkdir(base)
-    for path in path_list:
-        __path = Path(f"{base}/{path}")
-        if not os.path.exists(__path):
-            os.mkdir(__path)
-
 class DynamiCLI:
     def __init__(self, name: str) -> None:
         self.__name = name
-        self.__cc = ConfigControl(f"~/.dynami/config/{name}.json")
+        self.__base = Path("~/.dynami").expanduser()
+        self.check_config_path()
+        self.__config_path = Path(f"{self.__base}/config/{self.__name}.json").expanduser()
+
+    def check_config_path(self) -> None:
+        # If path is not existing, create it
+        path_list = ["config", "cache"]
+        __path = Path(self.__base).expanduser()
+        if not os.path.exists(__path):
+            os.mkdir(__path)
+            print("✅ Created config path!")
+        for path in path_list:
+            __path = Path(f"{self.__base}/{path}").expanduser()
+            if not os.path.exists(__path):
+                os.mkdir(__path)
 
     def set_attr_config(self, attribute: str, value: str) -> dict:
-        self.__cc.set(attribute=args.attribute, value=value)
-        self.__cc.save()
+        pass
 
     def create_config(self, provider: str) -> dict:
         config = {}
         config["name"] = self.__name
         config["provider"] = provider
         if config["provider"] not in available_providers:
-            raise Exception("Provider not available!")
+            raise Exception("❌ Provider not available!")
         key_type = input("Key: ")
         # Check if key is a path
         config["key"] = {}
@@ -58,26 +61,26 @@ class DynamiCLI:
             config["key"]["type"] = "token"
         config["key"]["value"] = key_type
         config["domain"] = input("Domain: ")
-        config["record"] = input("Record: ")
         config["type"] = input("Type: ")
         if config["type"] not in available_record_types:
-            raise Exception("Record type not available!")
+            raise Exception("❌ Record type not available!")
         self.__config = config
-        print(self.__config)
-        self.__cc.save(self.__config)
         
     def delete_config(self) -> None:
-        os.remove(self.__config_path)
+        if os.path.exists(self.__config_path):
+            os.remove(self.__config_path)
+            print("✅ Config successfully deleted!")
+        else:
+            print("❌ Config not found!")
 
     def save(self) -> None:
         with open(self.__config_path, "w") as f:
             f.write(json.dumps(self.__config))
             f.close()
+        print(f"✅ Config successfully saved! You can find it under ~/.dynami/config/{self.__name}.json")
 
 
 if __name__ == "__main__":
-    check_config_path("~/.dynami")
-
     parser = argparse.ArgumentParser()
     sub = parser.add_subparsers(title="Configuration", description="Configuration options", dest="sub")
 
@@ -88,7 +91,7 @@ if __name__ == "__main__":
     # Record: Create
     record_create = record_sub.add_parser("create", help="Create a new record")
     record_create_group = record_create.add_argument_group("record_create")
-    record_create_group.add_argument("-c", "--config", help="Name of configuration", type=str, dest="config", metavar="CONFIG", default="default")
+    record_create_group.add_argument("-n", "--name", help="Name of configuration", type=str, dest="config", metavar="CONFIG", default="default")
     record_create_group.add_argument("-r", "--record", help="Record name", type=str, dest="record_name", default="@", metavar="RECORD")
     record_create_group.add_argument("-t", "--type", help="Record type", type=str, dest="record_type", choices=available_record_types, default="A", metavar="TYPE")
     record_create_group.add_argument("-v", "--value", help="Record value", type=str, dest="record_value", default=None, metavar="VALUE")
@@ -100,18 +103,18 @@ if __name__ == "__main__":
     # Config: Create
     config_create = config_sub.add_parser("create", help="Create a new configuration")
     config_create_group = config_create.add_argument_group("config_create")
-    config_create_group.add_argument("-n", "--name", help="Name of the configuration", type=str, metavar="NAME", dest="config", default="default")
+    config_create_group.add_argument("-n", "--name", help="Name of the configuration", type=str, metavar="CONFIG", dest="config", default="default")
     config_create_group.add_argument("-p", "--provider", help="DNS Provider", type=str, metavar="PROVIDER", dest="provider", choices=available_providers)
 
     # Config: Delete
     config_delete = config_sub.add_parser("delete", help="Delete a configuration")
     config_delete_group = config_delete.add_argument_group("config_delete")
-    config_delete_group.add_argument("-n", "--name", help="Name of the configuration", type=str, metavar="NAME", dest="config_delete_name")
+    config_delete_group.add_argument("-n", "--name", help="Name of the configuration", type=str, metavar="CONFIG", dest="config")
 
     # Config: Set attribute
     config_set = config_sub.add_parser("set", help="Set an attribute of a configuration")
     config_set_group = config_set.add_argument_group("config_set")
-    config_set_group.add_argument("-n", "--name", help="Name of the configuration", type=str, metavar="NAME", dest="config")
+    config_set_group.add_argument("-n", "--name", help="Name of the configuration", type=str, metavar="CONFIG", dest="config")
     config_set_group.add_argument("-a", "--attribute", help="Attribute to set", type=str, metavar="ATTRIBUTE", dest="config_set_attribute")
     config_set_group.add_argument("-v", "--value", help="Value to set", type=str, metavar="VALUE", dest="config_set_value")
 
@@ -128,7 +131,8 @@ if __name__ == "__main__":
             match args.config_parser:
                 case "create":
                     cli.create_config(provider=args.provider)
+                    cli.save()
                 case "delete":
-                    print("config", "delete", args.config)
+                    cli.delete_config()
                 case "set":
                     print("config", "set", args.config, args.config_set_attribute, args.config_set_value)
